+++ .scripts/commit-orig.sh
@@ -50,11 +50,13 @@ fi
 
 if [[ "x${CI:-}" = "xtrue" ]]; then
     pushd "${ORIG_DIR}" > /dev/null
-    git config user.name 'github-actions[bot]'
-    git config user.email 'github-actions[bot]@users.noreply.github.com'
-    git checkout --orphan temp
-    git add .
-    git commit -m 'update orig' || true
-    git push --force origin HEAD:orig
+    if [[ $(git status --porcelain=v1 | wc -l) -gt 0 ]]; then
+        git config user.name 'github-actions[bot]'
+        git config user.email 'github-actions[bot]@users.noreply.github.com'
+        git checkout --orphan temp
+        git add .
+        git commit -m 'update orig' || true
+        git push --force origin HEAD:orig
+    fi
     popd > /dev/null
 fi
